Migrate initiate sso endpoint
