+++ server/service/handler.go
@@ -24,7 +24,6 @@ import (
 
 // FleetEndpoints is a collection of RPC endpoints implemented by the Fleet API.
 type FleetEndpoints struct {
-	InitiateSSO endpoint.Endpoint
 	CallbackSSO endpoint.Endpoint
 	SSOSettings endpoint.Endpoint
 }
@@ -32,14 +31,12 @@ type FleetEndpoints struct {
 // MakeFleetServerEndpoints creates the Fleet API endpoints.
 func MakeFleetServerEndpoints(svc fleet.Service, urlPrefix string, limitStore throttled.GCRAStore, logger kitlog.Logger) FleetEndpoints {
 	return FleetEndpoints{
-		InitiateSSO: logged(makeInitiateSSOEndpoint(svc)),
 		CallbackSSO: logged(makeCallbackSSOEndpoint(svc, urlPrefix)),
 		SSOSettings: logged(makeSSOSettingsEndpoint(svc)),
 	}
 }
 
 type fleetHandlers struct {
-	InitiateSSO http.Handler
 	CallbackSSO http.Handler
 	SettingsSSO http.Handler
 }
@@ -50,7 +47,6 @@ func makeKitHandlers(e FleetEndpoints, opts []kithttp.ServerOption) *fleetHandle
 		return kithttp.NewServer(e, decodeFn, encodeResponse, opts...)
 	}
 	return &fleetHandlers{
-		InitiateSSO: newServer(e.InitiateSSO, decodeInitiateSSORequest),
 		CallbackSSO: newServer(e.CallbackSSO, decodeCallbackSSORequest),
 		SettingsSSO: newServer(e.SSOSettings, decodeNoParamsRequest),
 	}
@@ -227,7 +223,6 @@ func addMetrics(r *mux.Router) {
 }
 
 func attachFleetAPIRoutes(r *mux.Router, h *fleetHandlers) {
-	r.Handle("/api/v1/fleet/sso", h.InitiateSSO).Methods("POST").Name("intiate_sso")
 	r.Handle("/api/v1/fleet/sso", h.SettingsSSO).Methods("GET").Name("sso_config")
 	r.Handle("/api/v1/fleet/sso/callback", h.CallbackSSO).Methods("POST").Name("callback_sso")
 }
@@ -398,6 +393,7 @@ func attachNewStyleFleetAPIRoutes(r *mux.Router, svc fleet.Service, logger kitlo
 	ne.GET("/api/_version_/fleet/invites/{token}", verifyInviteEndpoint, verifyInviteRequest{})
 	ne.POST("/api/_version_/fleet/reset_password", resetPasswordEndpoint, resetPasswordRequest{})
 	ne.POST("/api/_version_/fleet/logout", logoutEndpoint, nil)
+	ne.POST("/api/_version_/fleet/sso", initiateSSOEndpoint, initiateSSORequest{})
 
 	limiter := ratelimit.NewMiddleware(limitStore)
 	ne.
