+++ server/service/service_sessions.go
@@ -6,7 +6,6 @@ import (
 	"encoding/base64"
 	"encoding/xml"
 	"fmt"
-	"net/url"
 	"time"
 
 	"github.com/fleetdm/fleet/v4/server/contexts/ctxerr"
@@ -37,58 +36,6 @@ func (svc *Service) SSOSettings(ctx context.Context) (*fleet.SessionSSOSettings,
 		SSOEnabled:  appConfig.SSOSettings.EnableSSO,
 	}
 	return settings, nil
-}
-
-// InitiateSSO initiates a Single Sign-On flow for a request to visit the
-// protected URL identified by redirectURL. It returns the URL of the identity
-// provider to make a request to to proceed with the authentication via that
-// external service, and stores ephemeral session state to validate the
-// callback from the identity provider to finalize the SSO flow.
-func (svc *Service) InitiateSSO(ctx context.Context, redirectURL string) (string, error) {
-	// skipauth: User context does not yet exist. Unauthenticated users may
-	// initiate SSO.
-	svc.authz.SkipAuthorization(ctx)
-
-	logging.WithLevel(ctx, level.Info)
-
-	appConfig, err := svc.ds.AppConfig(ctx)
-	if err != nil {
-		return "", ctxerr.Wrap(ctx, err, "InitiateSSO getting app config")
-	}
-
-	metadata, err := svc.getMetadata(appConfig)
-	if err != nil {
-		return "", ctxerr.Wrap(ctx, err, "InitiateSSO getting metadata")
-	}
-
-	serverURL := appConfig.ServerSettings.ServerURL
-	settings := sso.Settings{
-		Metadata: metadata,
-		// Construct call back url to send to idp
-		AssertionConsumerServiceURL: serverURL + svc.config.Server.URLPrefix + "/api/v1/fleet/sso/callback",
-		SessionStore:                svc.ssoSessionStore,
-		OriginalURL:                 redirectURL,
-	}
-
-	// If issuer is not explicitly set, default to host name.
-	var issuer string
-	entityID := appConfig.SSOSettings.EntityID
-	if entityID == "" {
-		u, err := url.Parse(serverURL)
-		if err != nil {
-			return "", ctxerr.Wrap(ctx, err, "parse server url")
-		}
-		issuer = u.Hostname()
-	} else {
-		issuer = entityID
-	}
-
-	idpURL, err := sso.CreateAuthorizationRequest(&settings, issuer)
-	if err != nil {
-		return "", ctxerr.Wrap(ctx, err, "InitiateSSO creating authorization")
-	}
-
-	return idpURL, nil
 }
 
 func (svc *Service) getMetadata(config *fleet.AppConfig) (*sso.Metadata, error) {
