+++ server/service/transport_sessions.go
@@ -2,21 +2,11 @@ package service
 
 import (
 	"context"
-	"encoding/json"
 	"net/http"
 
 	"github.com/fleetdm/fleet/v4/server/contexts/ctxerr"
 	"github.com/fleetdm/fleet/v4/server/sso"
 )
-
-func decodeInitiateSSORequest(ctx context.Context, r *http.Request) (interface{}, error) {
-	var req initiateSSORequest
-	err := json.NewDecoder(r.Body).Decode(&req)
-	if err != nil {
-		return nil, err
-	}
-	return req, nil
-}
 
 func decodeCallbackSSORequest(ctx context.Context, r *http.Request) (interface{}, error) {
 	err := r.ParseForm()
